+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-# [0.0.2](https://github.com/Kozea/redux-api-unrest/compare/v0.0.1...v0.0.2)
+# [1.0.0](https://github.com/Kozea/redux-api-unrest/compare/v0.0.2...v1.0.0)
+
+* Settings support. (`namespace`, `elementDelimiter`, `modifierDelimiter`, `modifierValueDelimiter`).
+* Refactor internals a bit.
+
+### [0.0.2](https://github.com/Kozea/redux-api-unrest/compare/v0.0.1...v0.0.2)
 
 * Add mix function to mix different block and support modifiers on all of them.
 
-# [0.0.1](https://github.com/Kozea/redux-api-unrest/compare/...v0.0.1)
+### [0.0.1](https://github.com/Kozea/redux-api-unrest/compare/...v0.0.1)
 
 * First release.